always start with default wallpapers wal colors

+++ .config/scripts/background.sh
@@ -1,7 +1,14 @@
 #!/bin/bash
-wal -q -i "$1"
+
+if [ -z "$1" ]; then
+  wallpaper="~/.config/backgrounds/default.jpg"
+else
+  wallpaper="$1"
+  killall -SIGUSR2 waybar
+fi
+
+wal -q -i $wallpaper
 source "$HOME/.cache/wal/colors.sh"
-killall -SIGUSR2 waybar
 sass ~/.config/swaync/style.scss ~/.config/swaync/style.css
 swaync-client -rs
 rm ~/.config/swaync/style.css.map
